pyproject.toml
[project]
name = "furia"
version = "0.1.0"
description = ""
authors = [
    {name = "Ludwig S",email = "silvaludwig@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "aiohttp",
    "beautifulsoup4",
    "requests (>=2.32.3,<3.0.0)",
    "selenium"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
